extracted the lib adding mechanism
