--- feature-libs/misc/schematics/add-misc/schema.ts
@@ -1,5 +1,0 @@
-export interface Schema {
-  project: string;
-  lazy: boolean;
-  features: string[];
-}
