+++ feature-libs/misc/schematics/constants.ts
@@ -1,0 +1,1 @@
+export const STOREFINDER_SCSS_FILE_NAME = 'storefinder.scss';
